refactor(list-suppliers): simplify category-based supplier retrieval
- Replaced category slug lookup with category ID
- Added category existence validation
- Returned mapped CompanyOut responses
- Simplified supplier retrieval flow
