+++ core/application/src/main/java/com/arka/usecase/party/ListSuppliersByCategoryUseCase.java
@@ -1,12 +1,14 @@
-package com.arka.usecase.supplier;
+package com.arka.usecase.party;
 
+import com.arka.dto.out.CompanyOut;
+import com.arka.mapper.CompanyMapper;
+import com.arka.mapper.CompanyMapperImpl;
 import com.arka.model.Company;
 import com.arka.exceptions.NotFoundException;
 import com.arka.gateway.repository.product.ProductCategoryGateway;
 import com.arka.gateway.repository.party.SupplierGateway;
 import lombok.RequiredArgsConstructor;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RequiredArgsConstructor
@@ -14,21 +16,24 @@ public class ListSuppliersByCategoryUseCase {
 
     private final ProductCategoryGateway productCategoryGateway;
     private final SupplierGateway supplierGateway;
+    private final CompanyMapper companyMapper =
+            new CompanyMapperImpl();
 
-    public List<Company> execute(String categorySlug){
+    public List<CompanyOut> execute(Long productCategoryId) {
 
-        List<Company> foundCompanies = new ArrayList<>();
+        validateExistingCategory(productCategoryId);
 
-        productCategoryGateway
-               .findProductCategoryBySlug(categorySlug)
-               .ifPresentOrElse(productCategory -> {
+        List<Company> foundSuppliers =
+                supplierGateway.getSuppliersByProductCategoryId(productCategoryId);
 
-                   foundCompanies.addAll(supplierGateway
-                           .getSuppliersByProductCategoryId(productCategory.id()));
+        return foundSuppliers.stream()
+                .map(companyMapper::toOut)
+                .toList();
+    }
 
-               }, () -> { throw new NotFoundException("Category not found");});
-
-        return foundCompanies;
-
+    private void validateExistingCategory(Long categoryId){
+        productCategoryGateway.findById(categoryId).orElseThrow(() ->
+                new NotFoundException(String.format(
+                        "Category with id %d not found", categoryId)));
     }
 }
